requirements.txt
fastapi
uvicorn
httpx
playwright
pandas
pdfplumber
python-dotenv
google-generativeai